splitting once
